+++ src/SmartPresence/Services/Employees/Queries/EmployeeTimeOffByIdAndYearResponse.cs
@@ -1,7 +1,4 @@
 ﻿using SmartPresence.Services.Employees.Model;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace SmartPresence.Services.Employees.Queries
 {
@@ -11,9 +8,9 @@ namespace SmartPresence.Services.Employees.Queries
         public int HolidayAccrued { get; set; }
         public int HolidayUsed { get; set; }
         public int HolidayTotal { get; set; }
-        public int LeaveAccrued { get; set; }
-        public int LeaveUsed { get; set; }
-        public int LeaveTotal { get; set; }
+        public double LeaveAccrued { get; set; }
+        public double LeaveUsed { get; set; }
+        public double LeaveTotal { get; set; }
 
         public EmployeeTimeOffByIdAndYearResponse(EmployeeTimeOff employeeTimeOff)
         {
